change terminal lat long from decimal to int

+++ database/migrations/2023_09_27_184842_create_terminals.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->unsignedBigInteger('toda_id');
             $table->string('terminal_name');
             $table->text('terminal_address')->nullable();
-            $table->decimal('terminal_long', 10, 7)->nullable();
-            $table->decimal('terminal_lat', 10, 7)->nullable();
+            $table->integer('terminal_long')->nullable();
+            $table->integer('terminal_lat')->nullable();
             $table->timestamps(); // Adds created_at and updated_at columns
             $table->foreign('toda_id')->references('id')->on('todas');
         });
